% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addContextMenu.R
\name{addContextMenu}
\alias{addContextMenu}
\title{Add a context menu to a UI element}
\usage{
addContextMenu(
  ui_element_id,
  menu_items,
  itemPadding = "5px",
  itemFontSize = "14px",
  menuWidth = "auto"
)
}
\arguments{
\item{ui_element_id}{The ID of the UI element to attach the context menu to.}

\item{menu_items}{A data frame with menu items (name and id).}

\item{itemPadding}{Padding for context menu items.}

\item{itemFontSize}{Font size for context menu items.}

\item{menuWidth}{Width of the context menu.}
}
\description{
Add a context menu to a UI element
}
\examples{
\dontrun{
  library(shiny)

  ui <- fluidPage(
    titlePanel("Context Menu Example"),
    mainPanel(
      actionButton("myButton", "Right-click Me!")
    )
  )

  server <- function(input, output) {
    menu_items <- data.frame(name = c("Option 1", "Option 2", "Option 3"),
                             id = c("opt1", "opt2", "opt3"))

    addContextMenu("myButton", menu_items,
                   itemPadding = '5px',
                   itemFontSize = '14px',
                   menuWidth = 'auto')
  }

  shinyApp(ui, server)
}

}
