% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrTitle.R
\name{hrTitle}
\alias{hrTitle}
\title{Create a horizontal line with a title.}
\usage{
hrTitle(
  textToShow,
  position = "center",
  color = "grey",
  textColor = "black",
  verticalPosition = "top"
)
}
\arguments{
\item{textToShow}{Character string. The text to display as the title.}

\item{position}{Character string. One of "left", "center", or "right" indicating the horizontal alignment of the title. Default is "center".}

\item{color}{Character string. The color of the horizontal line. Default is "grey".}

\item{textColor}{Character string. The color of the title text. Default is "black".}

\item{verticalPosition}{Character string. One of "top" or "inline" indicating the vertical position of the title with respect to the line. "top" positions the text above the line, while "inline" centers the text with the line. Default is "top".}
}
\value{
A \code{shiny.tagList} containing the styled horizontal line and title.
}
\description{
This function generates a horizontal line with an optional title displayed above or inline. The title position (left, center, right) and colors can be customized.
}
\examples{
\dontrun{
# Create a Shiny app using hrTitle
library(shiny)

ui <- fluidPage(
  hrTitle("Centered Header"),
  hrTitle("Left Header", "left"),
  hrTitle("Inline Right Header", "right", verticalPosition = "inline")
)

server <- function(input, output, session) {}

shinyApp(ui, server)
}
}
