% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showCodeButtonServer.R
\name{showCodeButtonServer}
\alias{showCodeButtonServer}
\title{Server function to create a show/hide code button}
\usage{
showCodeButtonServer(input, output, session, generated_code)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{generated_code}{A function that returns the generated code.}
}
\value{
None
}
\description{
This function sets up the server-side logic for a Shiny app to display a
button that toggles the visibility of generated code.
}
