% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlePanel.R
\name{titlePanel}
\alias{titlePanel}
\title{titlePanel: A Custom Panel for Shiny UI with Various Styling Options}
\usage{
titlePanel(
  ...,
  title = NULL,
  bg_color = NULL,
  header_font = "Arial",
  header_weight = "normal",
  header_font_size = "18px",
  header_color = "grey",
  id = NULL,
  border_thickness = "1px",
  border_color = "lightgrey",
  border_type = "solid",
  margin_top = "20px"
)
}
\arguments{
\item{...}{Widgets to be included in the panel.}

\item{title}{An optional title for the panel. Default is NULL.}

\item{bg_color}{Background color for the panel. Default is transparent.}

\item{header_font}{Font for the title. Default is "Arial".}

\item{header_weight}{Weight for the title font. Default is "normal".}

\item{header_font_size}{Font size for the title. Default is "18px".}

\item{header_color}{Color for the title. Default is "grey".}

\item{id}{User-defined ID for the panel. If not provided, a unique ID will be generated.}

\item{border_thickness}{Thickness of the panel border. Default is "1px".}

\item{border_color}{Color of the panel border. Default is "lightgrey".}

\item{border_type}{Type of the panel border. Default is "solid".}

\item{margin_top}{Top margin for the panel. Default is "20px".}
}
\value{
A shiny tag list containing the custom panel and its styles.
}
\description{
The `titlePanel` function provides a way to create a styled panel
in Shiny applications. This custom panel can include widgets,
an optional title, and several style customizations.
}
\examples{
\dontrun{
titlePanel(
    shiny::sliderInput("slider1", "Slider Input:", min = 0, max = 100, value = 50),
    shiny::textInput("text1", "Text Input:"),
    title = "Custom Title",
    bg_color = "red",
    header_font = "Times New Roman",
    header_weight = "bold",
    header_font_size = "24px",
    header_color = "blue",
    id = "myCustomID",
    border_thickness = "2px",
    border_color = "blue",
    border_type = "dashed",
    margin_top = "15px"
)
}
}
