% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SideBySide.R
\name{SideBySide}
\alias{SideBySide}
\title{Side By Side Shiny Widgets}
\usage{
SideBySide(..., shrinkWidgets = FALSE)
}
\arguments{
\item{...}{Widgets to be displayed.}

\item{shrinkWidgets}{Logical. If TRUE, the widgets will be resized based on the number of widgets.}
}
\value{
Display widgets side by side.
}
\description{
Displays shiny widgets side by side.
}
