% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableLayout.R
\name{tableLayout}
\alias{tableLayout}
\title{Custom Table Layout for Shiny Widgets}
\usage{
tableLayout(
  ...,
  firstColWidth = NULL,
  fontSize = "16px",
  paddingBottom = "15px",
  isBold = TRUE
)
}
\arguments{
\item{...}{Shiny UI input elements.}

\item{firstColWidth}{The width of the first column (for the labels). If NULL, uses adaptive CSS.}

\item{fontSize}{The font size for the label text. Default is "16px".}

\item{paddingBottom}{The padding at the bottom of the label text. Default is "15px".}

\item{isBold}{Logical; If TRUE, the label text is displayed in bold. Default is TRUE.}
}
\description{
This function provides a custom table layout for arranging Shiny widgets with their associated labels.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    tableLayout(
      textInput(inputId = "text1", label = "First Input", value = "Example"),
      checkboxInput(inputId = "check1", label = "A Checkbox", value = TRUE)
    )
  )

  server <- function(input, output, session) {}

  shinyApp(ui, server)
}

}
